Move boss counting module into separate directory
