+++ ItemChanger.Silksong/SilksongHost.cs
@@ -4,6 +4,7 @@ using ItemChanger.Events;
 using ItemChanger.Logging;
 using ItemChanger.Modules;
 using ItemChanger.Silksong.Modules;
+using ItemChanger.Silksong.Modules.BossKillsCounter;
 using ItemChanger.Silksong.Modules.Obstacles;
 using ItemChanger.Silksong.Modules.YNBox;
 using ItemChanger.Silksong.Util;
